Added functionalities.
TODO: Fix the design so the squares are asymetrical as in the Figma design.

Fix the mapping of the answers to the participation methods.

Fix the design.

Fix the last letter on Scenarioubow.

Add search bar and functional searching capability.

Make it possible to swipe between pop ups for the methods so you can press on one method and then swipe on the pop up window to go to the next or previous method.